Remove document service
The document service originally became part of this repository because
it was used in a Mapzen demo/workshop some time ago as a more convenient
way to index documents for custom importers.

While it could be useful, it isn't currently supported and is not used
in this repository. By removing it from all the docker-compose.yml
files, we can reduce the size of downloaded Docker images.