docker: run chown on Elasticsearch directory
With Elasticsearch 5, proper permissions (by default for user 1000) must
be set on the data directory on the host system.

Users have consistently run the `pelias` helper script as users with a
different UID than 1000, or the one configured in `.env` (usually root),
so this change uses `chown` to set ownership of the Elasticsearch data
directory.

Connects https://github.com/pelias/docker/issues/33