Attempt to catch cases of incorrect datadir ownership
If the Elasticsearch datadir is not owned by the user specified in
$DOCKER_USER, Elasticsearch will fail to start.

Since https://github.com/pelias/docker/pull/55 the helper scripts
attempt to remedy this situation if the `pelias` script is run as root
(which is not recommended but is often done).

However, if the `pelias` script is not run as root and the permissions
are incorrect, this situation cannot be automatically fixed.

This code attempts to detect that case and recommend the proper command
to run (with sudo) and set proper directory permissions.

Connects https://github.com/pelias/docker/issues/31
Connects https://github.com/pelias/docker/issues/73