Update Elasticsearch 7 image to 7.5.1
While we've been testing ES7 support, Elastic has released a few new
versions. The latest is now 7.5.1.

This PR updates the images to use this new version.

Separately I have also built and pushed the image as
`pelias/elasticsearch:7.5.1`. The `pelias/elasticsearch:7.4.2` image is
still on Docker Hub for anyone who was using it.