Clarify that `ln` command may require sudo
While other `pelias` commands should not be run with `sudo`, `sudo` will
generally be required to create a symbolic link in
`/usr/local/bin/pelias` as described in our quickstart script.

This change adds `sudo` to that command and clarifies that it's the only
part of the quickstart script that should use `sudo`.

Connects https://github.com/pelias/docker/issues/214